+++ src/core/components/courses/ProfileFollowUp.vue
@@ -1,5 +1,5 @@
 <template>
-    <div>
+  <div>
     <p class="text-weight-bold">Progression des participants</p>
     <q-card>
       <ni-expanding-table :data="learners" :columns="columns" :pagination="pagination" :hide-bottom="false">
@@ -36,15 +36,16 @@
 </template>
 
 <script>
+import { mapGetters } from 'vuex';
 import Courses from '@api/Courses';
 import ExpandingTable from '@components/table/ExpandingTable';
 import Progress from '@components/CourseProgress';
 import { NotifyNegative } from '@components/popup/notify';
-import { eLearningCourseProfileMixin } from '@mixins/eLearningCourseProfileMixin';
+import { sortStrings, formatIdentity } from '@helpers/utils';
+import { E_LEARNING } from '@data/constants.js';
 
 export default {
   name: 'ProfileFollowUp',
-  mixins: [eLearningCourseProfileMixin],
   components: {
     'ni-expanding-table': ExpandingTable,
     'ni-progress': Progress,
@@ -52,17 +53,67 @@ export default {
   props: {
     profileId: { type: String, required: true },
   },
+  data () {
+    const isClientInterface = !/\/ad\//.test(this.$router.currentRoute.path);
+
+    return {
+      course: {},
+      tableLoading: false,
+      columns: [
+        {
+          name: 'name',
+          label: 'Nom',
+          field: 'identity',
+          format: value => (value ? value.fullName : ''),
+          align: 'left',
+          sortable: true,
+          sort: (a, b) => sortStrings(a.lastname, b.lastname),
+          style: 'width: 70%',
+        },
+        {
+          name: 'progress',
+          label: 'Progression',
+          field: 'progress',
+          align: 'center',
+          sortable: true,
+          style: 'min-width: 150px; width: 20%',
+        },
+        { name: 'expand', label: '', field: '' },
+      ],
+      learners: [],
+      pagination: { sortBy: 'name', ascending: true, page: 1, rowsPerPage: 15 },
+      E_LEARNING,
+      isClientInterface,
+    };
+  },
+  computed: {
+    ...mapGetters({ company: 'main/getCompany' }),
+  },
   async created () {
     await this.getLearnersList();
   },
   methods: {
     goToLearnerProfile (row) {
-      this.$router.push({ name: 'ni users learners info', params: { learnerId: row._id, defaultTab: 'courses' } });
+      const name = this.isClientInterface ? 'ni courses learners info' : 'ni users learners info';
+      this.$router.push({ name, params: { learnerId: row._id, defaultTab: 'courses' } });
+    },
+    formatRow (trainee) {
+      const formattedName = formatIdentity(trainee.identity, 'FL');
+
+      return {
+        _id: trainee._id,
+        identity: { ...trainee.identity, fullName: formattedName },
+        progress: trainee.progress,
+        steps: trainee.steps,
+      };
     },
     async getLearnersList () {
       try {
         this.tableLoading = true;
-        const course = await Courses.getFollowUp(this.profileId);
+        const course = await Courses.getFollowUp(
+          this.profileId,
+          this.isClientInterface ? { company: this.company._id } : null
+        );
 
         if (course) this.learners = Object.freeze(course.trainees.map(this.formatRow));
       } catch (e) {
