+++ src/modules/client/pages/ni/courses/ELearningCourseProfile.vue
@@ -1,105 +1,47 @@
 <template>
     <q-page class="client-background" padding>
       <ni-profile-header :title="courseName" />
-      <div>
-        <p class="text-weight-bold">Progression des participants</p>
-        <q-card>
-          <ni-expanding-table :data="learners" :columns="columns" :pagination="pagination" :hide-bottom="false">
-            <template #row="{ props }">
-              <q-td v-for="col in props.cols" :key="col.name" :props="props">
-                <template v-if="col.name === 'progress'">
-                  <ni-progress class="progress" :value="col.value" />
-                </template>
-                <template v-else-if="col.name === 'expand'">
-                  <q-icon :name="props.expand ? 'expand_less' : 'expand_more'" />
-                </template>
-                <template v-else>
-                  <div class="name" @click.stop="goToLearnerProfile(props.row)">{{ col.value }}</div>
-                </template>
-              </q-td>
-            </template>
-            <template #expanding-row="{ props }">
-              <q-td colspan="100%">
-                <div v-for="(step, stepIndex) in props.row.steps" :key="step._id" :props="props"
-                  class="q-ma-sm expanding-table-expanded-row">
-                  <div>
-                    <q-icon :name="step.type === E_LEARNING ? 'stay_current_portrait' : 'mdi-teach'" />
-                    {{ stepIndex + 1 }} - {{ step.name }}
-                  </div>
-                  <div class="expanding-table-progress-container">
-                    <ni-progress class="expanding-table-sub-progress" :value="step.progress" />
-                  </div>
-                </div>
-              </q-td>
-            </template>
-          </ni-expanding-table>
-        </q-card>
-      </div>
+      <ni-profile-follow-up :profile-id="courseId" />
     </q-page>
 </template>
 
 <script>
-import { mapGetters } from 'vuex';
+import { mapState } from 'vuex';
 import get from 'lodash/get';
-import Courses from '@api/Courses';
 import ProfileHeader from '@components/ProfileHeader';
-import ExpandingTable from '@components/table/ExpandingTable';
-import Progress from '@components/CourseProgress';
+import ProfileFollowUp from '@components/courses/ProfileFollowUp';
 import { NotifyNegative } from '@components/popup/notify';
-import { eLearningCourseProfileMixin } from '@mixins/eLearningCourseProfileMixin';
 
 export default {
   name: 'ELearningCourseProfile',
-  mixins: [eLearningCourseProfileMixin],
   components: {
     'ni-profile-header': ProfileHeader,
-    'ni-expanding-table': ExpandingTable,
-    'ni-progress': Progress,
+    'ni-profile-follow-up': ProfileFollowUp,
   },
   props: {
-    profileId: { type: String, required: true },
-  },
-  data () {
-    return {
-      course: {},
-    };
+    courseId: { type: String, required: true },
   },
   computed: {
-    ...mapGetters({ company: 'main/getCompany' }),
+    ...mapState('course', ['course']),
     courseName () {
       return get(this.course, 'subProgram.program.name');
     },
   },
   async created () {
-    await this.getLearnersList();
+    await this.refreshCourse();
   },
   methods: {
-    goToLearnerProfile (row) {
-      this.$router.push({ name: 'ni courses learners info', params: { learnerId: row._id, defaultTab: 'courses' } });
-    },
-    async getLearnersList () {
+    async refreshCourse () {
       try {
-        this.tableLoading = true;
-        this.course = await Courses.getFollowUp(this.profileId, { company: this.company._id });
-
-        if (this.course) this.learners = Object.freeze(this.course.trainees.map(this.formatRow));
+        await this.$store.dispatch('course/fetchCourse', { courseId: this.courseId });
       } catch (e) {
         console.error(e);
-        NotifyNegative('Erreur lors de la récupération des apprenants');
-        this.learners = [];
-      } finally {
-        this.tableLoading = false;
+        NotifyNegative('Erreur lors de la récupération de la formation.');
       }
     },
   },
+  beforeDestroy () {
+    this.$store.dispatch('course/resetCourse');
+  },
 };
 </script>
-
-<style lang="stylus" scoped>
-  .progress
-    width: 100%
-  .name
-    width: fit-content;
-    text-decoration: underline
-    color: $primary
-</style>
